+++ src/components/banner/Comment/Item/index.tsx
@@ -1,9 +1,9 @@
-import { postDeleteComment } from '@/apis/postDeleteComment';
 import { BannerComment } from '@/types/banner';
 import React from 'react';
 import styles from '../../../styles/components/banner/detail/BannerCommentItem.module.scss';
 import { FiClock } from 'react-icons/fi';
 import { RxPerson } from 'react-icons/rx';
+import { postDeleteComment } from '@/apis/comment';
 function CommentItem({
   bannerId,
   commentData,
