+++ src/components/banner/Comment/List/index.tsx
@@ -1,9 +1,8 @@
 import { BannerCommentPageableType } from '@/types/banner';
 import React, { useEffect, useState } from 'react';
-import CommentItem from './CommentItem';
-import { postComment } from '@/apis/postComment';
-import { getComments } from '@/apis/getComments';
 import styles from '../../../styles/components/banner/detail/BannerCommetList.module.scss';
+import { getComments, postComment } from '@/apis/comment';
+import CommentItem from '../Item';
 
 function CommentList({ id }: { id: number }) {
   const [text, setText] = useState('');
