+++ src/components/banner/Detail/Title/styles.module.scss
