+++ src/pages/banner/detail/[id].tsx
@@ -1,7 +1,6 @@
-import { getBannerItem } from '@/apis/getBannerItem';
-import CommentList from '@/components/Banner/detail/CommentList';
-import DetailButtons from '@/components/Banner/detail/DetailButtons';
-import DetailTitle from '@/components/Banner/detail/DetailTitle';
+import CommentList from '@/components/Banner/Comment/List';
+import DetailButtons from '@/components/Banner/Detail/Button';
+import DetailTitle from '@/components/Banner/Detail/Title';
 import { BannerItemType } from '@/types/banner';
 import {
   GetServerSideProps,
@@ -11,6 +10,7 @@ import {
 import Image from 'next/image';
 import { ParsedUrlQuery } from 'querystring';
 import styles from './styles.module.scss';
+import { getBannerItem } from '@/apis/banner';
 
 function InfoPage({ data }: { data: BannerItemType }) {
   const {
